Refactor context propagators to be carrier agnostic
