+++ src/Datadog.Trace/Propagation/DDSpanContextPropagator.cs
@@ -7,6 +7,9 @@ using Datadog.Trace.Logging;
 
 namespace Datadog.Trace.Propagation
 {
+    /// <summary>
+    /// Class that handles Datadog style context propagation.
+    /// </summary>
     internal class DDSpanContextPropagator : IPropagator
     {
         private const NumberStyles NumberStyles = System.Globalization.NumberStyles.Integer;
@@ -28,37 +31,6 @@ namespace Datadog.Trace.Propagation
         }
 
         public static DDSpanContextPropagator Instance { get; } = new DDSpanContextPropagator();
-
-        /// <summary>
-        /// Propagates the specified context by adding new headers to a <see cref="IHeadersCollection"/>.
-        /// This locks the sampling priority for <paramref name="context"/>.
-        /// </summary>
-        /// <param name="context">A <see cref="SpanContext"/> value that will be propagated into <paramref name="headers"/>.</param>
-        /// <param name="headers">A <see cref="IHeadersCollection"/> to add new headers to.</param>
-        public void Inject(SpanContext context, IHeadersCollection headers)
-        {
-            if (context == null) { throw new ArgumentNullException(nameof(context)); }
-
-            if (headers == null) { throw new ArgumentNullException(nameof(headers)); }
-
-            // lock sampling priority when span propagates.
-            context.TraceContext?.LockSamplingPriority();
-
-            headers.Set(DDHttpHeaderNames.TraceId, context.TraceId.ToString(InvariantCulture));
-            headers.Set(DDHttpHeaderNames.ParentId, context.SpanId.ToString(InvariantCulture));
-
-            // avoid writing origin header if not set, keeping the previous behavior.
-            if (context.Origin != null)
-            {
-                headers.Set(DDHttpHeaderNames.Origin, context.Origin);
-            }
-
-            var samplingPriority = (int?)(context.TraceContext?.SamplingPriority ?? context.SamplingPriority);
-
-            headers.Set(
-                DDHttpHeaderNames.SamplingPriority,
-                samplingPriority?.ToString(InvariantCulture));
-        }
 
         /// <summary>
         /// Propagates the specified context by adding new headers to a <see cref="IHeadersCollection"/>.
@@ -96,33 +68,6 @@ namespace Datadog.Trace.Propagation
         /// <summary>
         /// Extracts a <see cref="SpanContext"/> from the values found in the specified headers.
         /// </summary>
-        /// <param name="headers">The headers that contain the values to be extracted.</param>
-        /// <returns>A new <see cref="SpanContext"/> that contains the values obtained from <paramref name="headers"/>.</returns>
-        public SpanContext Extract(IHeadersCollection headers)
-        {
-            if (headers == null)
-            {
-                throw new ArgumentNullException(nameof(headers));
-            }
-
-            var traceId = ParseUInt64(headers, DDHttpHeaderNames.TraceId);
-
-            if (traceId == 0)
-            {
-                // a valid traceId is required to use distributed tracing
-                return null;
-            }
-
-            var parentId = ParseUInt64(headers, DDHttpHeaderNames.ParentId);
-            var samplingPriority = ParseSamplingPriority(headers, DDHttpHeaderNames.SamplingPriority);
-            var origin = headers.ParseString(DDHttpHeaderNames.Origin);
-
-            return new SpanContext(traceId, parentId, samplingPriority, null, origin);
-        }
-
-        /// <summary>
-        /// Extracts a <see cref="SpanContext"/> from the values found in the specified headers.
-        /// </summary>
         /// <param name="carrier">The headers that contain the values to be extracted.</param>
         /// <param name="getter">The function that can extract a list of values for a given header name.</param>
         /// <typeparam name="T">Type of header collection</typeparam>
@@ -143,39 +88,14 @@ namespace Datadog.Trace.Propagation
 
             var parentId = ParseUInt64(carrier, getter, DDHttpHeaderNames.ParentId);
             var samplingPriority = ParseSamplingPriority(carrier, getter, DDHttpHeaderNames.SamplingPriority);
-            var origin = ParseString(carrier, getter, DDHttpHeaderNames.Origin);
+            var origin = PropagationHelpers.ParseString(carrier, getter, DDHttpHeaderNames.Origin);
 
             return new SpanContext(traceId, parentId, samplingPriority, null, origin);
-        }
-
-        private static ulong ParseUInt64(IHeadersCollection headers, string headerName)
-        {
-            var headerValues = headers.GetValues(headerName);
-
-            var hasValue = false;
-
-            foreach (var headerValue in headerValues)
-            {
-                if (ulong.TryParse(headerValue, NumberStyles, InvariantCulture, out var result))
-                {
-                    return result;
-                }
-
-                hasValue = true;
-            }
-
-            if (hasValue)
-            {
-                Log.Warning("Could not parse {HeaderName} headers: {HeaderValues}", headerName, string.Join(",", headerValues));
-            }
-
-            return 0;
         }
 
         private static ulong ParseUInt64<T>(T carrier, Func<T, string, IEnumerable<string>> getter, string headerName)
         {
             var headerValues = getter(carrier, headerName);
-
             var hasValue = false;
 
             foreach (var headerValue in headerValues)
@@ -194,36 +114,6 @@ namespace Datadog.Trace.Propagation
             }
 
             return 0;
-        }
-
-        private static SamplingPriority? ParseSamplingPriority(IHeadersCollection headers, string headerName)
-        {
-            var headerValues = headers.GetValues(headerName);
-
-            var hasValue = false;
-
-            foreach (var headerValue in headerValues)
-            {
-                if (int.TryParse(headerValue, out var result))
-                {
-                    if (MinimumSamplingPriority <= result && result <= MaximumSamplingPriority)
-                    {
-                        return (SamplingPriority)result;
-                    }
-                }
-
-                hasValue = true;
-            }
-
-            if (hasValue)
-            {
-                Log.Warning(
-                    "Could not parse {HeaderName} headers: {HeaderValues}",
-                    headerName,
-                    string.Join(",", headerValues));
-            }
-
-            return default;
         }
 
         private static SamplingPriority? ParseSamplingPriority<T>(T carrier, Func<T, string, IEnumerable<string>> getter, string headerName)
@@ -254,21 +144,6 @@ namespace Datadog.Trace.Propagation
             }
 
             return default;
-        }
-
-        private static string ParseString<T>(T carrier, Func<T, string, IEnumerable<string>> getter, string headerName)
-        {
-            var headerValues = getter(carrier, headerName);
-
-            foreach (var headerValue in headerValues)
-            {
-                if (!string.IsNullOrEmpty(headerValue))
-                {
-                    return headerValue;
-                }
-            }
-
-            return null;
         }
     }
 }
