+++ src/Datadog.Trace/Propagation/IPropagator.cs
@@ -1,43 +1,27 @@
 using System;
 using System.Collections.Generic;
-using Datadog.Trace.Headers;
 
 namespace Datadog.Trace.Propagation
 {
     internal interface IPropagator
     {
         /// <summary>
-        /// Propagates the specified context by adding new headers to a <see cref="IHeadersCollection"/>.
+        /// Propagates the specified context by adding new headers to a carrier
         /// This locks the sampling priority for <paramref name="context"/>.
         /// </summary>
-        /// <param name="context">A <see cref="SpanContext"/> value that will be propagated into <paramref name="headers"/>.</param>
-        /// <param name="headers">A <see cref="IHeadersCollection"/> to add new headers to.</param>
-        void Inject(SpanContext context, IHeadersCollection headers);
-
-        /// <summary>
-        /// Propagates the specified context by adding new headers to a <see cref="IHeadersCollection"/>.
-        /// This locks the sampling priority for <paramref name="context"/>.
-        /// </summary>
-        /// <param name="context">A <see cref="SpanContext"/> value that will be propagated into <paramref name="carrier"/>.</param>
-        /// <param name="carrier">The headers to add to.</param>
+        /// <param name="context">A <see cref="SpanContext"/> value that will be propagated into <paramref name="carrier"/> instance.</param>
+        /// <param name="carrier">The headers metadata carrier to add to.</param>
         /// <param name="setter">The action that can set a header in the carrier.</param>
-        /// <typeparam name="T">Type of header collection</typeparam>
+        /// <typeparam name="T">Type of the headers metadata carrier</typeparam>
         void Inject<T>(SpanContext context, T carrier, Action<T, string, string> setter);
 
         /// <summary>
-        /// Extracts a <see cref="SpanContext"/> from the values found in the specified headers.
+        /// Extracts a <see cref="SpanContext"/> from the values found in the specified headers carrier.
         /// </summary>
-        /// <param name="headers">The headers that contain the values to be extracted.</param>
-        /// <returns>A new <see cref="SpanContext"/> that contains the values obtained from <paramref name="headers"/>.</returns>
-        SpanContext Extract(IHeadersCollection headers);
-
-        /// <summary>
-        /// Extracts a <see cref="SpanContext"/> from the values found in the specified headers.
-        /// </summary>
-        /// <param name="carrier">The headers that contain the values to be extracted.</param>
+        /// <param name="carrier">The headers metadata carrier to extract data from.</param>
         /// <param name="getter">The function that can extract a list of values for a given header name.</param>
-        /// <typeparam name="T">Type of header collection</typeparam>
-        /// <returns>A new <see cref="SpanContext"/> that contains the values obtained from <paramref name="carrier"/>.</returns>
+        /// <typeparam name="T">Type of the headers metadata carrier</typeparam>
+        /// <returns>A new <see cref="SpanContext"/> that contains the values obtained from <paramref name="carrier"/> instance.</returns>
         SpanContext Extract<T>(T carrier, Func<T, string, IEnumerable<string>> getter);
     }
 }
