+++ src/Datadog.Trace/Propagation/PropagationExtensions.cs
@@ -5,6 +5,16 @@ namespace Datadog.Trace.Propagation
 {
     internal static class PropagationExtensions
     {
+        public static void Inject(this IPropagator propagator, SpanContext context, IHeadersCollection headers)
+        {
+            propagator.Inject(context, headers, (carrier, header, value) => carrier.Set(header, value));
+        }
+
+        public static SpanContext Extract(this IPropagator propagator, IHeadersCollection headers)
+        {
+            return propagator.Extract(headers, (carrier, header) => carrier.GetValues(header));
+        }
+
         public static IEnumerable<KeyValuePair<string, string>> ExtractHeaderTags(this IHeadersCollection headers, IEnumerable<KeyValuePair<string, string>> headerToTagMap)
         {
             foreach (KeyValuePair<string, string> headerNameToTagName in headerToTagMap)
@@ -20,17 +30,7 @@ namespace Datadog.Trace.Propagation
 
         public static string ParseString(this IHeadersCollection headers, string headerName)
         {
-            var headerValues = headers.GetValues(headerName);
-
-            foreach (string headerValue in headerValues)
-            {
-                if (!string.IsNullOrEmpty(headerValue))
-                {
-                    return headerValue;
-                }
-            }
-
-            return null;
+            return PropagationHelpers.ParseString(headers, (carrier, header) => carrier.GetValues(header), headerName);
         }
     }
 }
